requirements.txt
# Core FastAPI and web framework dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
starlette>=0.27.0

# Data validation and serialization
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database and ORM
sqlalchemy>=2.0.23
alembic>=1.13.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Caching and message broker
redis>=5.0.1
celery>=5.3.4

# Vector database
qdrant-client>=1.7.0

# AI and ML frameworks
dspy-ai>=2.4.9
litellm>=1.17.0
openai>=1.6.0
anthropic>=0.8.0

# Document processing
docling>=2.36.0
docling-core>=1.0.0
docling-ibm-models>=1.0.0
python-docx>=1.1.0
PyPDF2>=3.0.1
PyMuPDF>=1.23.0
python-magic>=0.4.27
pillow>=10.0.0

# Text processing and NLP
sentence-transformers>=2.2.2
spacy>=3.7.0
nltk>=3.8.1
beautifulsoup4>=4.12.0
lxml>=4.9.0
html2text>=2020.1.16

# Web scraping and URL processing
requests>=2.31.0
aiohttp>=3.9.0
newspaper3k>=0.2.8
trafilatura>=1.6.0
readability-lxml>=0.8.1

# Monitoring and logging
prometheus-client>=0.19.0
python-json-logger>=2.0.7
structlog>=23.2.0

# System monitoring
psutil>=5.9.6

# Security and authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# HTTP client and utilities
httpx>=0.25.0
tenacity>=8.2.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0
factory-boy>=3.3.0

# Development tools
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0
pre-commit>=3.6.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.4.0

# Deployment and containerization
gunicorn>=21.2.0
docker>=6.1.0

# Additional utilities
click>=8.1.0
rich>=13.7.0
typer>=0.9.0
pyyaml>=6.0.1
jinja2>=3.1.0
python-dateutil>=2.8.0
pytz>=2023.3
